+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Added
 
 - Allow calling `login` command with a second argument to select WC contexts.
+- Add `-clientcert` suffix to WC client certificate contexts created by the `login` command.
+- Use `CertificateAuthorityData` to store CA data and ensure that `CertificateAuthority` is not set when manipulating the kubeconfig in the `login` command.
 
 ## [2.16.0] - 2022-07-01